Make INVALID_ANGLE a float rather than a double, since the variables we're assigning (and comparing) it to are float.
git-svn-id: http://svn/svn/hcr/branches/pmc730_daemon@1173 2e3f476d-dfe7-4101-bf40-6ec35b1080dd

+++ src/hcrgui/HcrGuiMainWindow.cpp
@@ -16,7 +16,7 @@
 #include <logx/Logging.h>
 
 // Invalid angle used to erase display of reflector position
-static const double INVALID_ANGLE = -999.9;
+static const float INVALID_ANGLE = -999.9;
 
 LOGGING("HcrGuiMainWindow")
